// src/app/generated/blueprint.api/model/steamfitterTask.ts
/**
 * Blueprint API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SteamfitterIntegrationType } from './steamfitterIntegrationType';
import { SteamfitterTaskTrigger } from './steamfitterTaskTrigger';
import { ScenarioEvent } from './scenarioEvent';
import { SteamfitterTaskAction } from './steamfitterTaskAction';


export interface SteamfitterTask { 
    dateCreated?: Date;
    dateModified?: Date | null;
    createdBy?: string;
    modifiedBy?: string | null;
    id?: string;
    scenarioEventId?: string;
    scenarioEvent?: ScenarioEvent;
    taskType?: SteamfitterIntegrationType;
    name?: string | null;
    description?: string | null;
    action?: SteamfitterTaskAction;
    vmMask?: string | null;
    apiUrl?: string | null;
    actionParameters?: { [key: string]: string; } | null;
    expectedOutput?: string | null;
    expirationSeconds?: number;
    delaySeconds?: number;
    intervalSeconds?: number;
    iterations?: number;
    triggerCondition?: SteamfitterTaskTrigger;
    userExecutable?: boolean;
    repeatable?: boolean;
}